Add option to compile unverified code

+++ build.sh
@@ -1,8 +1,20 @@
-#!/bin/sh
+#!/bin/bash
 
-set -xe
+set -e
 
-agda2hs ./src/Lam/Agda/Lam/Export.agda
-mv ./src/Lam/Agda/Lam/*.hs ./src/Lam
-rm ./src/Lam/Export.hs
-stack build
+if [[ $* == *--help* ]]
+then
+    echo "USAGE: \`./build.sh --verified\` for verified code or \`./build.sh --unverified\` for unverified (more features)"
+elif [[ $* == *--verified* ]]
+then
+    agda2hs ./src/Lam/Agda/Lam/Export.agda
+    mv ./src/Lam/Agda/Lam/*.hs ./src/Lam
+    rm ./src/Lam/Export.hs
+elif [[ $* == *--unverified* ]]
+then
+    cp ./src/Lam/Unverified/*.hs ./src/Lam
+else
+    echo "USAGE: \`./build.sh --verified\` for verified code or \`./build.sh --unverified\` for unverified (more features)"
+fi
+
+
